DO NOT MERGE
Drafted up a possible public-facing version. Do not merge, this is just for internal sharing.

Possibly add some sort of graph/graphic that moves with the data up top. Maybe semantic mapping, maybe

It has a password section that we can change regularly, or kill entirely.

I also left a section specifically in here that will print a line on top of the data that can say something along the lines of "data from xyz, please cite as...." Right now it just has filer text.